requirements.txt
streamlit
PyPDF2~=3.0.1
langchain
langchain-google-genai~=2.0.6
google-generativeai
faiss-cpu
python-docx~=1.1.2
beautifulsoup4~=4.12.3
requests~=2.32.3
protobuf~=5.29.1